- Moved controllers to Controller folder - Fixed serialization issue (circular reference)
